Fix int to uint comparison warning, QList.size() always returns a positive or zero number, but int is still used at least until Qt7.x  by size() (resp. count()) methods, it was used by Qt developers in order to avoid ugly code in Qt codebase.

+++ src/Common/gui/chords/ChordProgression.cpp
@@ -19,7 +19,7 @@ bool ChordProgression::canBeUsed(int bpi) const
 {
     auto dividers = bpiUtils::getBpiDividers(bpi);
     for (uint divider : dividers) {
-        if (divider == measures.size())
+        if (divider == (uint) measures.size())
             return true;
     }
